+++ service/twitter/doc.go
@@ -1,0 +1,9 @@
+// Copyright 2021 Wayback Archiver. All rights reserved.
+// Use of this source code is governed by the GNU GPL v3
+// license that can be found in the LICENSE file.
+
+/*
+Package twitter implements the twitter daemon service.
+*/
+
+package twitter // import "github.com/wabarc/wayback/service/twitter"